+++ indico/migrations/versions/20231106_1517_0acf26d68434_add_speakers_can_submit_to_events.py
@@ -1,4 +1,4 @@
-"""add_speakers_can_submit_to_events
+"""Add speakers_can_submit to events table
 
 Revision ID: 0acf26d68434
 Revises: 31b699664893
@@ -7,8 +7,6 @@ Create Date: 2023-11-06 15:17:38.522399
 
 import sqlalchemy as sa
 from alembic import op
-
-from indico.modules.events.models.settings import EventSetting
 
 
 # revision identifiers, used by Alembic.
@@ -21,7 +19,6 @@ depends_on = None
 def upgrade():
     op.add_column('events', sa.Column('speakers_can_submit', sa.Boolean(), nullable=False,
                   server_default='false'), schema='events')
-    op.alter_column('events', sa.Column('speakers_can_submit', server_default=None), schema='events')
     op.execute('''
         UPDATE events.events ev
         SET speakers_can_submit = true
@@ -38,19 +35,15 @@ def upgrade():
         es.module = 'subcontributions' AND
         es.name = 'speakers_can_submit';
     ''')
+    op.alter_column('events', sa.Column('speakers_can_submit', server_default='true'), schema='events')
 
 
 def downgrade():
     conn = op.get_bind()
-    events_speakers_can_submit = conn.execute('''
-        SELECT id FROM events.events ev
+    conn.execute('''
+        INSERT INTO events.settings(module, name, event_id, value)
+        SELECT 'subcontributions', 'speakers_can_submit', id, 'true'::jsonb FROM events.events ev
         WHERE
         ev.speakers_can_submit;
     ''')
-    values = []
-    for id in events_speakers_can_submit:
-        values.append((id, 'subcontributions', 'speakers_can_submit', 'true'))
-    stmt = sa.insert(EventSetting).values(
-        [{'name': name, 'module': module, 'event_id': event_id, 'value': value} for event_id, module, name, value in values])
-    op.execute(stmt)
     op.drop_column('events', 'speakers_can_submit', schema='events')