+++ indico/modules/events/management/controllers/protection.py
@@ -18,7 +18,6 @@ from indico.modules.categories.models.roles import CategoryRole
 from indico.modules.categories.util import serialize_category_role
 from indico.modules.core.controllers import PrincipalsMixin
 from indico.modules.events import Event
-from indico.modules.events.contributions import subcontribution_settings
 from indico.modules.events.controllers.base import RHAuthenticatedEventBase
 from indico.modules.events.management.controllers.base import RHManageEventBase
 from indico.modules.events.management.forms import EventProtectionForm
@@ -105,7 +104,7 @@ class RHEventProtection(RHManageEventBase):
         permissions = [[serialize_principal(p.principal), list(get_principal_permissions(p, Event))]
                        for p in self.event.acl_entries]
         permissions = [item for item in permissions if item[1]]
-        subcontrib_speakers = subcontribution_settings.get(self.event, 'speakers_can_submit')
+        subcontrib_speakers = self.event.speakers_can_submit
 
         return dict({'protection_mode': self.event.protection_mode, 'registration_managers': registration_managers,
                      'access_key': self.event.access_key, 'visibility': self.event.visibility,
@@ -120,7 +119,7 @@ class RHEventProtection(RHManageEventBase):
         update_session_coordinator_privs(self.event, data)
 
     def _update_subcontrib_settings(self, form):
-        subcontribution_settings.set(self.event, 'speakers_can_submit', form.subcontrib_speakers.data)
+        self.event.speakers_can_submit = form.subcontrib_speakers.data
 
 
 class RHPermissionsDialog(RH):